# openapi/oas3-combined.yaml
openapi: 3.0.3
info:
  title: Uranus API
  description: |
    API for the Uranus / kulturbytes platform.

    Public event API for Uranus.

    The event endpoints are public and support filtering, pagination,
    calendar views, summaries, and GeoJSON venue output.
  version: '1.0'
servers:
  - url: https://api.kulturbytes.de
  - url: http://localhost:9090
...
paths:
  /api/event/{eventUuid}:
    get:
      tags:
        - Events
      summary: Get an event
      description: Returns a single event. Selects the earliest date on or after today in the
        server local timezone (including dates earlier today). If no such date exists, selects
        the latest past date. The selected date is returned in date and the remaining dates
        in further_dates. If the event has no dates, both fields are omitted. Public access
        is limited to released, cancelled, deferred and rescheduled events. If the request context
        contains a user with edit, delete, release or insights permissions for the event organizer,
        draft and review events are also allowed.
      operationId: getEvent
      parameters:
        - name: eventUuid
          in: path
          required: true
          description: Event UUID.
          schema:
            type: string
            format: uuid
        - name: lang
          in: query
          required: false
          description: Language used for localized type, genre, license and accessibility labels.
          schema:
            type: string
            default: en
            example: en
      responses:
        '200':
          description: Event successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEventResponse200'
              example:
                service: Uranus API
                api_version: '1.0'
                response_type: get-event
                status: 200
                timestamp: '2026-09-21T10:00:00Z'
                metadata:
                  response_time_ms: 12
                data:
                  uuid: 01996f4a-7c00-7000-8000-000000000001
                  release_status: released
                  content_language: en
                  title: Evening concert
                  org_uuid: 01996f4a-7c00-7000-8000-000000000002
                  org_name: Kulturbytes
                  org_logos: {}
                  date_match: false
                  date:
                    uuid: 01996f4a-7c00-7000-8000-000000000003
                    slug: '202609301930'
                    event_uuid: 01996f4a-7c00-7000-8000-000000000001
                    release_status: released
                    start_date: '2026-09-30'
                    start_time: '19:30'
                    venue_uuid: null
                    venue_logos: {}
        '500':
          description: Internal server error. This handler also returns 500 when the event does
            not exist or is not visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEventResponse500'
  /api/event/{eventUuid}/date/{dateIdentifier}:
    get:
      tags:
        - Events
      summary: Get an event by date
      description: Returns an event with the requested date selected. If the identifier cannot
        be resolved or the date does not belong to the event, uses the default date selection.
        Selects the earliest date on or after today in the server local timezone (including
        dates earlier today). If no such date exists, selects the latest past date. The selected
        date is returned in date and the remaining dates in further_dates. If the event has
        no dates, both fields are omitted. Public access is limited to released, cancelled,
        deferred and rescheduled events. If the request context contains a user with edit, delete,
        release or insights permissions for the event organizer, draft and review events are
        also allowed.
      operationId: getEventByDate
      parameters:
        - name: eventUuid
          in: path
          required: true
          description: Event UUID.
          schema:
            type: string
            format: uuid
        - name: dateIdentifier
          in: path
          required: true
          description: Event date UUIDv7 or a date slug in YYYYMMDDHHMM format. An unresolved
            identifier falls back to the default date selection.
          schema:
            type: string
            example: '202609301930'
        - name: lang
          in: query
          required: false
          description: Language used for localized type, genre, license and accessibility labels.
          schema:
            type: string
            default: en
            example: en
      responses:
        '200':
          description: Event successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEventByDateResponse200'
              example:
                service: Uranus API
                api_version: '1.0'
                response_type: get-event-by-date
                status: 200
                timestamp: '2026-09-21T10:00:00Z'
                metadata:
                  response_time_ms: 12
                  event_uuid: 01996f4a-7c00-7000-8000-000000000001
                  date_identifier: '202609301930'
                  lang: en
                  event_date_count: 1
                data:
                  uuid: 01996f4a-7c00-7000-8000-000000000001
                  release_status: released
                  content_language: en
                  title: Evening concert
                  org_uuid: 01996f4a-7c00-7000-8000-000000000002
                  org_name: Kulturbytes
                  org_logos: {}
                  date_match: true
                  date:
                    uuid: 01996f4a-7c00-7000-8000-000000000003
                    slug: '202609301930'
                    event_uuid: 01996f4a-7c00-7000-8000-000000000001
                    release_status: released
                    start_date: '2026-09-30'
                    start_time: '19:30'
                    venue_uuid: null
                    venue_logos: {}
        '400':
          description: A required event or date identifier is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEventByDateResponse400'
        '404':
          description: Event not found or not visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEventByDateResponse404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEventByDateResponse500'
  /api/event/category-lookup:
    get:
      tags:
        - Events
      summary: Get event category lookup
      description: >-
        Returns event categories grouped by language and category ID. The
        optional lang parameter can contain a comma-separated list of ISO 639-1
        language codes.
      operationId: getEventCategoryLookup

      parameters:
        - name: lang
          in: query
          required: false
          description: >-
            Comma-separated list of ISO 639-1 language codes used to filter the
            returned categories, e.g. "de,en,da". If omitted or empty, all
            available languages are returned.
          schema:
            type: string
            example: de,en,da

      responses:
        '200':
          description: Event category lookup successfully retrieved.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EventCategoryLookupResponseEnvelope200'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventCategoryLookup'
                    required:
                      - data

              example:
                service: Uranus API
                api_version: '1.0'
                response_type: get-event-category-lookup
                status: 200
                timestamp: '2026-09-20T13:30:00Z'
                data:
                  da:
                    '1':
                      name: Kultur
                    '2':
                      name: Musik
                    '3':
                      name: Teater
                  de:
                    '1':
                      name: Kultur
                    '2':
                      name: Musik
                    '3':
                      name: Theater
                  en:
                    '1':
                      name: Culture
                    '2':
                      name: Music
                    '3':
                      name: Theatre

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCategoryLookupResponseEnvelope500'
...
  /api/choosable-countries:
    get:
      operationId: getChoosableCountries
      summary: Get choosable countries
      description: |
        Returns all countries available for selection.

        Country names are localized according to the requested
        ISO 639-1 language code.

        If no countries are found for the requested language, the endpoint
        still returns HTTP 200 with an empty data array.

      parameters:
        - name: lang
          in: query
          required: false
          description: ISO 639-1 language code. Defaults to English.
          schema:
            type: string
            minLength: 2
            maxLength: 2
            default: en
            example: en

      responses:

        '200':
          description: Countries found, or no countries available for the requested language.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ChoosableCountriesResponseEnvelope200'
                  - type: object
                    properties:
                      data:
                        type: array
                        description: List of available countries.
                        items:
                          $ref: '#/components/schemas/Country'
                    required:
                      - data

              example:
                service: Uranus API
                api_version: '1.0'
                response_type: get-choosable-countries
                status: 200
                timestamp: '2026-09-20T09:48:48Z'
                metadata:
                  country_count: 249
                  language: en
                  response_time_ms: 95
                data:
                  - country_code: AFG
                    country_name: Afghanistan
                  - country_code: ALB
                    country_name: Albania
                  - country_code: DZA
                    country_name: Algeria
                  - country_code: ASM
                    country_name: American Samoa
                  - country_code: AND
                    country_name: Andorra

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ChoosableCountriesResponseEnvelope500'
                  - type: object
                    properties:
                      data:
                        nullable: true

              example:
                service: Uranus API
                api_version: '1.0'
                response_type: get-choosable-countries
                status: 500
                timestamp: '2026-09-20T09:48:48Z'
                metadata:
                  language: en
                  response_time_ms: 95
                data: null
                message: Internal server error
    servers:
    - url: https://api.kulturbytes.de
    - url: https://localhost:9090
...
  /api/events/filter:
    post:
      tags: [Events]
      summary: Get events using JSON payload for filtering
      description: |
        Returns events matching the supplied JSON filter object.

        Unknown JSON properties are ignored by the current implementation.
        Array-valued properties are represented as JSON arrays.
      operationId: filterEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventFilterRequest'
      responses:
        '200':
          description: Events returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    servers:
    - url: https://api.kulturbytes.de
      description: Production API
    - url: http://localhost:9090
      description: Local API
...
components:
  schemas:
...
    GetEventResponse200:
      allOf:
        - $ref: '#/components/schemas/EventDetailsResponseEnvelope'
        - type: object
          properties:
            response_type:
              type: string
              enum:
                - get-event
            status:
              type: integer
              enum:
                - 200
            data:
              $ref: '#/components/schemas/EventDetails'
          required:
            - data
    GetEventResponse500:
      allOf:
        - $ref: '#/components/schemas/EventDetailsResponseEnvelope'
        - type: object
          properties:
            response_type:
              type: string
              enum:
                - get-event
            status:
              type: integer
              enum:
                - 500
          required:
            - message
...
    EventCategoryLookupResponseEnvelope200:
      allOf:
        - $ref: '#/components/schemas/EventCategoryLookupResponseEnvelope'
        - type: object
          properties:
            status:
              type: integer
              enum:
                - 200
              example: 200
    EventCategoryLookupResponseEnvelope500:
      allOf:
        - $ref: '#/components/schemas/EventCategoryLookupResponseEnvelope'
        - type: object
          properties:
            status:
              type: integer
              enum:
                - 500
              example: 500
...